just gotta fix a few bugs now
